³É¹¦Êä³öÑéÖ¤Âë

+++ config/chensong.go
@@ -5,8 +5,8 @@ import "time"
 type ChenSongConfig struct {
 	ApiUrl         string        `mapstructure:"api_url"`
 	ApiToken       string        `mapstructure:"api_token"`
-	ActivatedQQ    int           `mapstructure:"activated_qq"`
-	ActivatedGroup int           `mapstructure:"activated_group"`
+	ActivatedQQ    int64         `mapstructure:"activated_qq"`
+	ActivatedGroup int64         `mapstructure:"activated_group"`
 	BindMaxTries   int           `mapstructure:"bind_max_tries"`
 	BindTimeout    time.Duration `mapstructure:"bind_timeout"`
 	ReceiveToken   string        `mapstructure:"receive_token"`
